+++ docs/release-notes.rst
@@ -22,6 +22,16 @@ Design Studio Release Notes
 
 The Design Studio Composition Engine is deployed within a Docker image in the Docker registry.  TOSCAGenerator Client is uploaded in Nexus repository as a jar file.
 
+0.0.31-SNAPSHOT, 2018-06-12
+---------------------------
+* ACUMOS-1003 : Update CE Modify Node API for Collator (BE).
+* ACUMOS-1004 : Update CE Delete Node API for Collator (BE).
+* ACUMOS-1005 : Update CE Delete Link API for Collator (BE).
+* ACUMOS-1055 : Update Modify Node API for Splitter (BE).
+* ACUMOS-1056 : Update Add Link API for Splitter/Collator (BE).
+* ACUMOS-1057 : Update Delete Link API for Splitter (BE).
+
+
 0.0.30-SNAPSHOT, 2018-06-06
 ---------------------------
 * ACUMOS-971 : Deploy button active for invalid solution, BluePrint File Changes.